refactor: removed InvalidUserError.
