+++ src/error/index.ts
@@ -1,6 +1,5 @@
 export * from "./invalid-credentials.error";
 export * from "./invalid-refresh-token.error";
 export * from "./invalid-reset-password-request.error";
-export * from "./invalid-user.error";
 export * from "./user-exists.error";
 export * from "./user-not-found.error";
