--- src/error/invalid-user.error.ts
@@ -1,6 +1,0 @@
-export class InvalidUserError extends Error {
-  constructor() {
-    super();
-    this.message = "Invalid User.";
-  }
-}
